feat: add ability to pass attrs to demo blocks
This commit adds the ability to pass attrs to demo blocks by adding an `attrs` prop to the `GenerateOptions` interface. This allows for more flexibility in how the demo component is rendered, as different attributes can be passed to the `highlight` function depending on the use case. The demo container and demo component functions have been updated to take this new prop into account, and the `sfcHtml` variable has been replaced with `sfcHtml` which takes both `sfcTsHtml` and `sfcJsHtml` into account. The `api-examples.md` file has also been updated to show usage of the new `attrs` prop.